Refactoring stage 0.5. Removed most of the boost dependencies, refactored the hierarchy of the pruning classes, introduced the optimization process (pre or post traninig), replaced the boost CLI parser with a home made parser, moved the application logic from the main method to a driver

+++ include/learning/forests/rankboost.h
@@ -90,7 +90,7 @@ private:
 /// The Journal of machine learning research, 4, 933-969.
 class Rankboost : public LTR_Algorithm {
 public:
-    Rankboost(unsigned int max_wr);
+    Rankboost(size_t max_wr);
 
     Rankboost(const boost::property_tree::ptree &info_ptree,
               const boost::property_tree::ptree &model_ptree);
@@ -138,8 +138,8 @@ private:
     unsigned int*** SDF = NULL;
     Score* training_scores = NULL;
     Score* validation_scores = NULL;
-    unsigned int T;
-    unsigned int best_T;
+    size_t T;
+    size_t best_T;
     bool go_parallel;
     char const* omp_schedule;
     WeakRanker** weak_rankers = NULL;
